fix the modal
