+++ packages/apps/docs/src/components/Search/Search.tsx
@@ -1,0 +1,58 @@
+import { SearchResults } from './components/SearchResults';
+
+import { useSearch } from '@/hooks';
+import { useSemanticSearch } from '@/hooks/useSearch/useSemanticSearch';
+import React, { FC, useEffect, useState } from 'react';
+
+interface IProps {
+  query?: string;
+  hasScroll?: boolean;
+}
+
+export const Search: FC<IProps> = ({ query, hasScroll }) => {
+  const [tabName, setTabName] = useState<string | undefined>();
+  const {
+    outputStream,
+    handleSubmit: handleSearchSubmit,
+    conversation,
+    error,
+  } = useSearch();
+  const {
+    results: semanticResults,
+    error: semanticError,
+    isLoading: semanticIsLoading,
+    handleSubmit: handleSemanticSubmit,
+  } = useSemanticSearch();
+
+  useEffect(() => {
+    if (query !== undefined) {
+      if (tabName === 'qa') {
+        // eslint-disable-next-line @typescript-eslint/no-floating-promises
+        handleSearchSubmit(query);
+      }
+      if (tabName === 'docs') {
+        handleSemanticSubmit(query);
+      }
+    }
+  }, [query, tabName, handleSemanticSubmit, handleSearchSubmit]);
+
+  const onTabSelect = (tabName: string): void => {
+    setTabName(tabName);
+  };
+
+  return (
+    <section>
+      <SearchResults
+        semanticResults={semanticResults}
+        semanticError={semanticError}
+        semanticIsLoading={semanticIsLoading}
+        conversation={conversation}
+        outputStream={outputStream}
+        query={query}
+        error={error}
+        hasScroll={hasScroll}
+        onTabSelect={onTabSelect}
+      />
+    </section>
+  );
+};
