+++ packages/apps/docs/src/components/SearchModal/styles.ts
@@ -2,4 +2,5 @@ import { styled, StyledComponent } from '@kadena/react-components';
 
 export const Wrapper: StyledComponent<'div'> = styled('div', {
   width: '100%',
+  minHeight: '60vh',
 });
